fixed plotting

+++ ros_ws/src/SLAM/landmark_based_slam.py
@@ -246,8 +246,8 @@ class EKFSlamNode(Node):
         cv2.circle(img, (px, py), 5, (255, 0, 0), -1)
         arrow_length = 20 # pixels
         arrow_length_m = arrow_length * self.PLOT_METERS / self.PLOT_PIXELS
-        arrow_dx = int(arrow_length_m * math.cos(th))
-        arrow_dy = int(arrow_length_m * math.sin(th))
+        arrow_dx = arrow_length_m * math.cos(th)
+        arrow_dy = arrow_length_m * math.sin(th)
         px2, py2 = self._world_to_px(x + arrow_dx, y + arrow_dy)
         cv2.arrowedLine(img, (px, py), (px2, py2), (255, 0, 0), 2)
